Refactor Docker environment
In particular, the Dockerfile has been refactored to be straight even
the gems will be installed as root. The docker-compose has been
refactored as well to allow an environment to develop, which is more
manageable than a raw command.

The directory dedicated to the site has been moved to a stardard one.