Add math support and refactor dates
The math support has been added using both KaTeX and MathJax. The engine
can be chosen in the config file.

The concept of the dates throught the site has been changed. In order to
adhere to the original theme, only the date variable can be set in the
front matter. It can be used as a rolling date, instead, the one in the
file name is the starting one.