Add emoji support and reading time
The support to emoji has been added with jemoji. The reading time has
been added with jekyll-time-to-read.

It has been start the usage of @use in Sass file since @import is
deprecated and will be removed. A line in Gemfile has been added in
order to correctly handle google-protobuf since it was preventing the
building of the site.