Add rich content
The rich content support has been added. The only rich content support
not added is the Instagram one. Since in next releases of the original
Anatole it has been not included, it is excluded here.

It has been fixed an error on authors of posts.

In the .gitignore, Gemfile.lock has been added in order to avoid
unintended commits.